cambios para que funcione cd - y coja bien las rutas
