+++ minishell.h
@@ -116,5 +116,6 @@ void custom_env(t_content *cont, int i);
 void custom_cd(t_content *cont, int i);
 void update_environment_new(t_content *cont, int i, char *new_pwd);
 void update_environment_old(t_content *cont, int i, char *old_pwd);
+char * get_the_oldpwd(t_content *cont, int i);
 
 #endif
